requirements.txt
annotated-types==0.7.0
asgiref==3.10.0
cffi==2.0.0
contextlib2==21.6.0
cryptography==46.0.3
Django==5.2.8
django-cors-headers==4.9.0
django-ninja==1.4.5
django-ninja-extra==0.30.2
django-ninja-jwt==5.4.0
injector==0.22.0
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
sqlparse==0.5.3
typing-inspection==0.4.2
typing_extensions==4.15.0
